Pushing few more changes

+++ urban-data-comparator/src/main/java/com/yantriks/urbandatacomparator/model/UrbanCsvOutputData.java
@@ -19,7 +19,7 @@ public class UrbanCsvOutputData {
 
     private String reservationStatus;
 
-    private int reservationResponseCode;
+    private long reservationResponseCode;
 
     private String error;
